backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
alembic>=1.7.0
pydantic>=1.8.0
python-dotenv>=0.19.0
httpx>=0.20.0
redis>=4.0.0
celery>=5.0.0
PyJWT>=2.0.0
schedule>=1.1.0
python-multipart>=0.0.5
aiofiles>=0.7.0